Extract primary test helper plugin functions to dedicated file
